Add support for hex, octal, binary integers (TOML v1.0)
This also adds additional tests for unsigned integers and big.Int.

When trying to unmarshal a negative integer into uintXX, the error
message now says that the problem is the integer being negative.

+++ parse.go
@@ -124,7 +124,7 @@ func (p *tomlParser) SetTime(begin, end int) {
 	}
 }
 
-func (p *tomlParser) SetFloat64(begin, end int) {
+func (p *tomlParser) SetFloat(begin, end int) {
 	p.val = &ast.Float{
 		Position: ast.Position{Begin: begin, End: end},
 		Data:     p.buffer[begin:end],
@@ -132,7 +132,7 @@ func (p *tomlParser) SetFloat64(begin, end int) {
 	}
 }
 
-func (p *tomlParser) SetInt64(begin, end int) {
+func (p *tomlParser) SetInteger(begin, end int) {
 	p.val = &ast.Integer{
 		Position: ast.Position{Begin: begin, End: end},
 		Data:     p.buffer[begin:end],
